# apps/docs/src/openapi/tradr-api.json
{
...
  "info": {
    "description": "HTTP API for Tradr — the open-source trading journal with an AI advisor.\n\nThis reference is generated from the `@swagger` JSDoc blocks that live next to each route in the API source, so it always matches the endpoints the app actually serves. All routes are mounted under `/api`. The same API backs both the hosted app (`api.tradr.cloud`) and a self-hosted instance (reached through the web container at `/api`). Most endpoints require an authenticated session cookie.",
    "title": "Tradr API",
    "version": "0.5.3"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/api/brokerages": {
      "get": {
        "description": "Authed. Returns the user's own brokerages plus the built-in system ones, each with its fee schedule. A brokerage is a name and a fee schedule that accounts and positions reference so commissions are computed consistently. Tradr does not connect to any broker; nothing here holds credentials.\n",
        "responses": {
          "200": {
            "description": "The brokerages visible to this user."
          },
          "401": {
            "description": "No valid session."
          }
        },
        "summary": "List brokerages.",
        "tags": [
          "Brokerages"
        ]
      },
      "post": {
        "description": "Authed. Names are unique per user. The new brokerage starts with a zeroed fee schedule; set the rates with a follow-up PUT.\n",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "name": {
                    "maxLength": 100,
                    "minLength": 1,
                    "type": "string"
                  },
                  "notes": {
                    "maxLength": 10000,
                    "nullable": true,
                    "type": "string"
                  }
                },
                "required": [
                  "name"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "The created brokerage."
          },
          "400": {
            "description": "Validation error."
          },
          "409": {
            "description": "A brokerage with this name already exists."
          }
        },
        "summary": "Create a brokerage.",
        "tags": [
          "Brokerages"
        ]
      }
    },
    "/api/brokerages/{id}": {
      "delete": {
        "description": "Authed. Refused while any account still references the brokerage — the 409 names them, so you can reassign those accounts first. System brokerages cannot be deleted.\n",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted."
          },
          "403": {
            "description": "System brokerages cannot be deleted."
          },
          "404": {
            "description": "No such brokerage for this user."
          },
          "409": {
            "description": "Still assigned to one or more accounts",
            "which are named in the message.": null
          }
        },
        "summary": "Delete a brokerage.",
        "tags": [
          "Brokerages"
        ]
      },
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The brokerage",
            "including its fee schedule.": null
          },
          "404": {
            "description": "No such brokerage for this user."
          }
        },
        "summary": "Get one brokerage.",
        "tags": [
          "Brokerages"
        ]
      },
      "put": {
        "description": "Authed. Every field is optional; `feeSchedule` is merged, so you can send a single rate. Fee values are non-negative decimal strings. System brokerages are read-only.\n",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "feeSchedule": {
                    "description": "Any subset of the fee fields. Each is a non-negative decimal string.",
                    "properties": {
                      "optionsMaxPerFill": {
                        "type": "string"
                      },
                      "optionsMinPerFill": {
                        "type": "string"
                      },
                      "optionsPerContractCommission": {
                        "type": "string"
                      },
                      "optionsPerContractExchangeFee": {
                        "type": "string"
                      },
                      "stockMaxPerFill": {
                        "type": "string"
                      },
                      "stockMinPerFill": {
                        "type": "string"
                      },
                      "stockPerShareCommission": {
                        "type": "string"
                      }
                    },
                    "type": "object"
                  },
                  "name": {
                    "maxLength": 100,
                    "minLength": 1,
                    "type": "string"
                  },
                  "notes": {
                    "maxLength": 10000,
                    "nullable": true,
                    "type": "string"
                  }
                },
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The updated brokerage."
          },
          "400": {
            "description": "Validation error."
          },
          "403": {
            "description": "System brokerages cannot be modified."
          },
          "404": {
            "description": "No such brokerage for this user."
          },
          "409": {
            "description": "A brokerage with this name already exists."
          }
        },
        "summary": "Edit a brokerage or its fee schedule.",
        "tags": [
          "Brokerages"
        ]
      }
    },
    "/api/brokerages/{id}/duplicate": {
      "post": {
        "description": "Authed. Copies a brokerage and its whole fee schedule under a new name, which is how you take a system brokerage as the starting point for your own rates. The copy belongs to the user and is fully editable.\n",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "The new copy."
          },
          "404": {
            "description": "No such brokerage for this user."
          },
          "409": {
            "description": "The generated name collides with an existing brokerage."
          }
        },
        "summary": "Duplicate a brokerage.",
        "tags": [
          "Brokerages"
        ]
      }
    },
    "/api/brokerages/{id}/position-count": {
      "get": {
        "description": "Authed. Used to warn before a destructive edit.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "count": {
                      "type": "integer"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "The count."
          },
          "404": {
            "description": "No such brokerage for this user."
          }
        },
        "summary": "Count the positions booked against a brokerage.",
        "tags": [
          "Brokerages"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Hosted (tradr.cloud)",
      "url": "https://api.tradr.cloud"
    },
    {
      "description": "Local development (API PORT default)",
      "url": "http://localhost:3100"
    }
  ],
...
}